refactor: move all storage hanlders to storage.service
