+++ src/services/profile.service.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { getToken } from "services/auth.service";
+import { getToken } from "services/storage.service";
 
 let AUTH_URL = process.env.REACT_APP_API_URL;
 axios.defaults.baseURL = AUTH_URL;
@@ -13,17 +13,4 @@ export const getProfileInfo = () => {
         id: AUTH_ID,
         auth_key: AUTH_KEY
     }).then(response => response)
-}
-
-export const getProfile = () => {
-    return JSON.parse(sessionStorage.getItem("SWOB_USER"));
-}
-
-export const setProfile = (profile) => {
-    removeProfile()
-    sessionStorage.setItem('SWOB_USER', JSON.stringify(profile));
-}
-
-export const removeProfile = () => {
-    sessionStorage.removeItem("SWOB_USER");
 }
